Flask Server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.57.4
anthropic==0.42.0
groq==0.11.0
chromadb==0.5.23
sqlalchemy==2.0.25
Pillow==10.2.0
PyPDF2==3.0.1
python-docx==1.1.0
pytesseract==0.3.10
markdown==3.5.2
numpy>=1.22.0,<2.0.0
httpx==0.27.2

